+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "How time flies".
 - Tests for "Minesweeper level generator".
 - Tests for "In stereo".
+- Tests for "Hooch clash".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".